Issue #4587: Moved resources for descendant token check test
