+++ src/test/java/com/puppycrawl/tools/checkstyle/checks/DescendantTokenCheckTest.java
@@ -36,7 +36,10 @@ import com.puppycrawl.tools.checkstyle.utils.CommonUtils;
 public class DescendantTokenCheckTest extends BaseCheckTestSupport {
     @Override
     protected String getPath(String filename) throws IOException {
-        return super.getPath("checks" + File.separator + filename);
+        return super.getPath("checks" + File.separator
+                + "misc" + File.separator
+                + "descendanttoken" + File.separator
+                + filename);
     }
 
     @Test
@@ -45,7 +48,7 @@ public class DescendantTokenCheckTest extends BaseCheckTestSupport {
         final DefaultConfiguration checkConfig =
             createCheckConfig(DescendantTokenCheck.class);
         final String[] expected = CommonUtils.EMPTY_STRING_ARRAY;
-        verify(checkConfig, getPath("InputIllegalTokens.java"), expected);
+        verify(checkConfig, getPath("InputDescendantTokenIllegalTokens.java"), expected);
     }
 
     @Test
@@ -59,7 +62,7 @@ public class DescendantTokenCheckTest extends BaseCheckTestSupport {
         final String[] expected = {
             "17:12: " + getCheckMessage(MSG_KEY_MAX, 1, 0, "LITERAL_NATIVE", "LITERAL_NATIVE"),
         };
-        verify(checkConfig, getPath("InputIllegalTokens.java"), expected);
+        verify(checkConfig, getPath("InputDescendantTokenIllegalTokens.java"), expected);
     }
 
     @Test
@@ -74,7 +77,7 @@ public class DescendantTokenCheckTest extends BaseCheckTestSupport {
         final String[] expected = {
             "17:12: Using 'native' is not allowed.",
         };
-        verify(checkConfig, getPath("InputIllegalTokens.java"), expected);
+        verify(checkConfig, getPath("InputDescendantTokenIllegalTokens.java"), expected);
     }
 
     @Test
@@ -88,7 +91,7 @@ public class DescendantTokenCheckTest extends BaseCheckTestSupport {
         final String[] expected = {
             "8:9: " + getCheckMessage(MSG_KEY_MIN, 1, 2, "LITERAL_SWITCH", "LITERAL_DEFAULT"),
         };
-        verify(checkConfig, getPath("InputIllegalTokens.java"), expected);
+        verify(checkConfig, getPath("InputDescendantTokenIllegalTokens.java"), expected);
     }
 
     @Test
@@ -101,7 +104,7 @@ public class DescendantTokenCheckTest extends BaseCheckTestSupport {
         checkConfig.addAttribute("maximumNumber", "0");
         checkConfig.addAttribute("minimumDepth", "3");
         final String[] expected = CommonUtils.EMPTY_STRING_ARRAY;
-        verify(checkConfig, getPath("InputIllegalTokens.java"), expected);
+        verify(checkConfig, getPath("InputDescendantTokenIllegalTokens.java"), expected);
     }
 
     @Test
@@ -114,7 +117,7 @@ public class DescendantTokenCheckTest extends BaseCheckTestSupport {
         checkConfig.addAttribute("maximumNumber", "0");
         checkConfig.addAttribute("maximumDepth", "1");
         final String[] expected = CommonUtils.EMPTY_STRING_ARRAY;
-        verify(checkConfig, getPath("InputIllegalTokens.java"), expected);
+        verify(checkConfig, getPath("InputDescendantTokenIllegalTokens.java"), expected);
     }
 
     @Test
@@ -147,7 +150,7 @@ public class DescendantTokenCheckTest extends BaseCheckTestSupport {
             "75:10: Empty statement.",
         };
 
-        verify(checkConfig, getPath("InputEmptyStatement.java"), expected);
+        verify(checkConfig, getPath("InputDescendantTokenEmptyStatement.java"), expected);
     }
 
     @Test
@@ -164,7 +167,7 @@ public class DescendantTokenCheckTest extends BaseCheckTestSupport {
             "17:9: switch without \"default\" clause.",
         };
 
-        verify(checkConfig, getPath("InputMissingSwitchDefault.java"), expected);
+        verify(checkConfig, getPath("InputDescendantTokenMissingSwitchDefault.java"), expected);
     }
 
     @Test
@@ -183,7 +186,7 @@ public class DescendantTokenCheckTest extends BaseCheckTestSupport {
             "12:20: Literal Strings should be compared using equals(), not '=='.",
             "17:22: Literal Strings should be compared using equals(), not '=='.",
         };
-        verify(checkConfig, getPath("InputStringLiteralEquality.java"), expected);
+        verify(checkConfig, getPath("InputDescendantTokenStringLiteralEquality.java"), expected);
     }
 
     @Test
@@ -201,7 +204,7 @@ public class DescendantTokenCheckTest extends BaseCheckTestSupport {
             "11:18: Using 'POST_DEC' is not allowed.",
             "12:18: Using 'POST_INC' is not allowed.",
         };
-        verify(checkConfig, getPath("InputIllegalTokens.java"), expected);
+        verify(checkConfig, getPath("InputDescendantTokenIllegalTokens.java"), expected);
     }
 
     @Test
@@ -217,7 +220,7 @@ public class DescendantTokenCheckTest extends BaseCheckTestSupport {
         final String[] expected = {
             "17:12: Using 'LITERAL_NATIVE' is not allowed.",
         };
-        verify(checkConfig, getPath("InputIllegalTokens.java"), expected);
+        verify(checkConfig, getPath("InputDescendantTokenIllegalTokens.java"), expected);
     }
 
     @Test
@@ -234,7 +237,7 @@ public class DescendantTokenCheckTest extends BaseCheckTestSupport {
             "15:11: Return from catch is not allowed.",
         };
 
-        verify(checkConfig, getPath("InputReturnFromCatch.java"), expected);
+        verify(checkConfig, getPath("InputDescendantTokenReturnFromCatch.java"), expected);
     }
 
     @Test
@@ -251,7 +254,7 @@ public class DescendantTokenCheckTest extends BaseCheckTestSupport {
             "15:11: Return from finally is not allowed.",
         };
 
-        verify(checkConfig, getPath("InputReturnFromFinally.java"), expected);
+        verify(checkConfig, getPath("InputDescendantTokenReturnFromFinally.java"), expected);
     }
 
     @Test
@@ -265,7 +268,7 @@ public class DescendantTokenCheckTest extends BaseCheckTestSupport {
 
         final String[] expected = CommonUtils.EMPTY_STRING_ARRAY;
 
-        verify(checkConfig, getPath("InputReturnFromFinally.java"), expected);
+        verify(checkConfig, getPath("InputDescendantTokenReturnFromFinally.java"), expected);
     }
 
     @Test
@@ -285,7 +288,7 @@ public class DescendantTokenCheckTest extends BaseCheckTestSupport {
             "23:51: this cannot be null.",
         };
 
-        verify(checkConfig, getPath("InputReturnFromFinally.java"), expected);
+        verify(checkConfig, getPath("InputDescendantTokenReturnFromFinally.java"), expected);
     }
 
     @Test
@@ -304,7 +307,7 @@ public class DescendantTokenCheckTest extends BaseCheckTestSupport {
             "23:51: " + getCheckMessage(MSG_KEY_SUM_MAX, 2, 1, "NOT_EQUAL"),
         };
 
-        verify(checkConfig, getPath("InputReturnFromFinally.java"), expected);
+        verify(checkConfig, getPath("InputDescendantTokenReturnFromFinally.java"), expected);
     }
 
     @Test
@@ -325,7 +328,7 @@ public class DescendantTokenCheckTest extends BaseCheckTestSupport {
             "24:77: " + getCheckMessage(MSG_KEY_SUM_MIN, 1, 3, "EQUAL"),
         };
 
-        verify(checkConfig, getPath("InputReturnFromFinally.java"), expected);
+        verify(checkConfig, getPath("InputDescendantTokenReturnFromFinally.java"), expected);
     }
 
     @Test
@@ -347,6 +350,6 @@ public class DescendantTokenCheckTest extends BaseCheckTestSupport {
             "24:77: custom message",
         };
 
-        verify(checkConfig, getPath("InputReturnFromFinally.java"), expected);
+        verify(checkConfig, getPath("InputDescendantTokenReturnFromFinally.java"), expected);
     }
 }
