+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/misc/descendanttoken/InputDescendantTokenIllegalTokens.java
@@ -1,0 +1,18 @@
+package com.puppycrawl.tools.checkstyle.checks.misc.descendanttoken;
+
+public class InputDescendantTokenIllegalTokens
+{
+    public void methodWithPreviouslyIllegalTokens()
+    {
+        int i = 0;
+        switch (i)
+        {
+            default:
+                i--;
+                i++;
+                break;
+        }
+    }
+    
+    public native void nativeMethod();
+}
